+++ module/ldbc-query-builder/src/main/scala/ldbc/query/builder/syntax/CommandSyntax.scala
@@ -6,43 +6,12 @@
 
 package ldbc.query.builder.syntax
 
-import cats.data.Kleisli
-import cats.syntax.all.*
-
-import cats.effect.Temporal
-
-import ldbc.sql.{ Statement, ResultSet }
 import ldbc.dsl.*
 import ldbc.query.builder.statement.Command
 
-trait CommandSyntax[F[_]: Temporal]:
+trait CommandSyntax[F[_]]:
 
   extension (command: Command)
-    def update: Executor[F, Int] =
-      Executor.Impl[F, Int](
-        command.statement,
-        command.params,
-        connection =>
-          for
-            prepareStatement <- connection.prepareStatement(command.statement)
-            result <- command.params.zipWithIndex.traverse {
-                        case (param, index) => param.bind[F](prepareStatement, index + 1)
-                      } >> prepareStatement.executeUpdate() <* prepareStatement.close()
-          yield result
-      )
+    def update: Executor[F, Int]
 
-    def returning[T <: String | Int | Long](using reader: ResultSetReader[F, T]): Executor[F, T] =
-      given Kleisli[F, ResultSet[F], T] = Kleisli(resultSet => reader.read(resultSet, 1))
-
-      Executor.Impl[F, T](
-        command.statement,
-        command.params,
-        connection =>
-          for
-            prepareStatement <- connection.prepareStatement(command.statement, Statement.RETURN_GENERATED_KEYS)
-            resultSet <- command.params.zipWithIndex.traverse {
-                           case (param, index) => param.bind[F](prepareStatement, index + 1)
-                         } >> prepareStatement.executeUpdate() >> prepareStatement.getGeneratedKeys()
-            result <- summon[ResultSetConsumer[F, T]].consume(resultSet) <* prepareStatement.close()
-          yield result
-      )
+    def returning[T <: String | Int | Long](using reader: ResultSetReader[F, T]): Executor[F, T]
